+++ src/main/java/kr/or/cola/backend/todo/dto/TodoSaveRequestDto.java
@@ -1,0 +1,4 @@
+package kr.or.cola.backend.todo.dto;
+
+public class TodoSaveRequestDto {
+}
