+++ src/main/java/kr/or/cola/backend/todo/item/dto/ItemsResponseDto.java
@@ -1,7 +1,9 @@
-package kr.or.cola.backend.todo.folder.dto;
+package kr.or.cola.backend.todo.item.dto;
 
 import kr.or.cola.backend.todo.item.domain.Item;
+import lombok.Getter;
 
+@Getter
 public class ItemsResponseDto {
 
     private Long itemId;
